Start removing process_fn generators
